+++ simulacao.h
@@ -1,6 +1,6 @@
 #ifndef SIMULACAO_H
 #define SIMULACAO_H
 
-int nova_simulacao(int* cromossomo, unsigned long* seed);
+int nova_simulacao(int* cromossomo, unsigned* seed);
 
 #endif // SIMULACAO_H
